Move e2e tests into subfolder
