+++ cmd/ocw/e2e/testdata/basic/hello_world.yaml
