+++ cmd/ocw/e2e/testdata/outputs/output_passing.yaml
