+++ cmd/ocw/e2e/testdata/switch/switch_no_default.yaml
