+++ cmd/ocw/e2e/testdata/templates/template_workflow.yaml
